backend/requirements.txt
# Django
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0

# Database
psycopg2-binary>=2.9.9

# Utils
python-dotenv>=1.0.0
Pillow>=10.0.0
